feat: move assets from observable to assets.py
