+++ src/cala/assets.py
@@ -1,29 +1,184 @@
-from cala.models import Footprints, Traces, PixStat, CompStat, Overlap, Residual, Movie
+from copy import deepcopy
+from typing import ClassVar
+
+import xarray as xr
+from pydantic import BaseModel, PrivateAttr, field_validator
+
+from cala.models.axis import Coords, Dims
+from cala.models.checks import is_non_negative
+from cala.models.entity import Entity, Group
 
 
-def footprints() -> Footprints:
-    return Footprints()
+class Asset(BaseModel):
+    array: xr.DataArray | None = None
+    _entity: ClassVar[Entity]
+
+    class Config:
+        arbitrary_types_allowed = True
+        validate_assignment = True
+
+    def __eq__(self, other: "Asset") -> bool:
+        return self.array.equals(other.array)
+
+    @classmethod
+    def entity(cls) -> Entity:
+        return cls._entity
+
+    @field_validator("array", mode="after")
+    @classmethod
+    def validate_array_schema(cls, value: xr.DataArray) -> xr.DataArray:
+        value.validate.against_schema(cls._entity.model)
+
+        return value
 
 
-def traces() -> Traces:
-    return Traces()
+class Footprint(Asset):
+    _entity: ClassVar[Entity] = PrivateAttr(
+        Entity(
+            name="footprint",
+            dims=(Dims.width.value, Dims.height.value),
+            dtype=float,
+            checks=[is_non_negative],
+        )
+    )
 
 
-def pixel_stats() -> PixStat:
-    return PixStat()
+class Trace(Asset):
+    _entity: ClassVar[Entity] = PrivateAttr(
+        Entity(name="trace", dims=(Dims.frame.value,), dtype=float, checks=[is_non_negative])
+    )
 
 
-def comp_stats() -> CompStat:
-    return CompStat()
+class Frame(Asset):
+    _entity: ClassVar[Entity] = PrivateAttr(
+        Entity(
+            name="frame",
+            dims=(Dims.width.value, Dims.height.value),
+            dtype=float,
+            checks=[is_non_negative],
+        )
+    )
 
 
-def overlaps() -> Overlap:
-    return Overlap()
+class Footprints(Asset):
+    _entity: ClassVar[Entity] = PrivateAttr(
+        Group(
+            name="footprint-group",
+            member=Footprint.entity(),
+            group_by=Dims.component,
+            checks=[is_non_negative],
+        )
+    )
 
 
-def residuals() -> Residual:
-    return Residual()
+class Traces(Asset):
+    """
+    traces(path: str | Path | None = None, peek_size: int = 999)
+
+    Traces(array=array, path=path) -> saves to zarr (should be set in this asset, and leave
+    untouched in nodes.)
+    Traces.array -> loads from zarr
+
+    _array: xr.DataArray
+
+    @property
+    def array():
+        return (
+            xr.open_zarr(self.store_path)
+            .isel({AXIS.frames_dim: slice(-self.peek_size, None)})
+            .to_dataarray()
+            .isel({"variable": 0})  # not sure why it automatically makes this coordinate
+            .reset_coords("variable", drop=True)
+        )
+
+    @array.setter
+    def setter():
+        value.to_zarr(self.store_path, mode="w")  # need to make sure it can overwrite
+    """
+
+    _entity: ClassVar[Entity] = PrivateAttr(
+        Group(
+            name="trace-group",
+            member=Trace.entity(),
+            group_by=Dims.component,
+            checks=[is_non_negative],
+        )
+    )
 
 
-def buffer() -> Movie:
-    return Movie()
+class Movie(Asset):
+    _entity: ClassVar[Entity] = PrivateAttr(
+        Group(
+            name="movie",
+            member=Frame.entity(),
+            group_by=Dims.frame.value,
+            checks=[is_non_negative],
+        )
+    )
+
+
+class PopSnap(Asset):
+    """
+    A snapshot of a population trait.
+
+    Mainly used for Traces that only has one frame.
+    """
+
+    _entity: ClassVar[Entity] = PrivateAttr(
+        Entity(
+            name="pop-snap",
+            dims=(Dims.component.value,),
+            dtype=float,
+            coords=[Coords.frame.value, Coords.timestamp.value],
+        )
+    )
+
+
+comp_dims = (Dims.component.value, deepcopy(Dims.component.value))
+comp_dims[1].name += "'"
+for coord in comp_dims[1].coords:
+    coord.name += "'"
+
+
+class CompStats(Asset):
+    _entity: ClassVar[Entity] = PrivateAttr(
+        Entity(
+            name="comp-stat",
+            dims=comp_dims,
+            dtype=float,
+            checks=[],
+        )
+    )
+
+
+class PixStats(Asset):
+    _entity: ClassVar[Entity] = PrivateAttr(
+        Entity(
+            name="pix-stat",
+            dims=(Dims.width.value, Dims.height.value, Dims.component.value),
+            dtype=float,
+            checks=[],
+        )
+    )
+
+
+class Overlaps(Asset):
+    _entity: ClassVar[Entity] = PrivateAttr(
+        Entity(
+            name="overlap",
+            dims=comp_dims,
+            dtype=bool,
+            checks=[],
+        )
+    )
+
+
+class Residual(Asset):
+    _entity: ClassVar[Entity] = PrivateAttr(
+        Group(
+            name="frame",
+            member=Frame.entity(),
+            group_by=Dims.frame.value,
+            checks=[is_non_negative],
+        )
+    )
