+++ src/cala/models/__init__.py
@@ -1,33 +1,4 @@
 from .axis import AXIS
-from .observable import (
-    CompStat,
-    Footprint,
-    Footprints,
-    Frame,
-    Movie,
-    Overlap,
-    PixStat,
-    PopSnap,
-    Residual,
-    Trace,
-    Traces,
-)
-from .params import Params
 from .spec import Pipe
 
-__all__ = [
-    AXIS,
-    Pipe,
-    Footprint,
-    Footprints,
-    Trace,
-    Traces,
-    Frame,
-    Movie,
-    Params,
-    CompStat,
-    PixStat,
-    Overlap,
-    Residual,
-    PopSnap,
-]
+__all__ = [AXIS, Pipe]
